Remove automatic build behaviour of Program constructors
This closes Bug 10552.
The default values for the build booleans in the Program
constructors are removed. Also, the constructor without a
build boolean is removed as this builds by default.
The functioanl behaviour of creating a program for the
default Context with a kernal string is already implemented
by the Program (STRING_CLASS, bool, cl_int) constructor.


git-svn-id: https://cvs.khronos.org/svn/repos/OpenCL/branches/CLHPP@22425 e7fa87d3-cd2b-0410-9028-fcbf551c1848