Buffer constructor with iterators and specified context added. CommandQueue constructor to return queue to first device in context added. cl::copy commands to enqueue blocking copy for iterators on queue for buffer <-> host copies.
This closes bug # 10549.
The new Buffer constructor relies on the other two additions.
The new (explicit) CommandQueue constructor accepts a context and returns
a queue to the first device in the context.
The cl::copy enqueues a blocking copy for iterators on the specified command
queue - the original versions call this new one with CommandQueue::getDefault().
The buffer constructor is a blocking call.
Tests added to the test suite for all these methods.



git-svn-id: https://cvs.khronos.org/svn/repos/OpenCL/branches/CLHPP@22451 e7fa87d3-cd2b-0410-9028-fcbf551c1848