Update to C++ wrapper of clCreateContextFromType method
The constructor is updated to allow a more successful
aquisition of a context on a non-Apple platform.
Specifically, on non-Apple, clCreateContextFromType
requires the cl_context_properties array to be not null.
In the current implementation, only the first platform
is checked for a device of the specified type. This
update continues checking all the platforms found for a
device of the specified type.

Closes bug 10556.


git-svn-id: https://cvs.khronos.org/svn/repos/OpenCL/branches/CLHPP@22761 e7fa87d3-cd2b-0410-9028-fcbf551c1848