Fix failing tests and enable testing in Travis
Since we build with CL_HPP_MINIMUM_OPENCL_VERSION defined to 100,
the bindings are querying the platform to decide whether to use
clCreateCommandQueue or clCreateCommandQueueWithProperties. We
thus need to make the stub function for clGetPlatformInfo report
2.0 when building with CL_HPP_TARGET_OPENCL_VERSION >= 200 to
mirror the queue creation function check.

Also allow calls to clCreateContextFromType with no properties on
Apple platforms to mirror the logic in the bindings.

Contributes to #62.

Signed-off-by: Kevin Petit <kevin.petit@arm.com>