Make it possible to use the repository as a submodule directly
The new cl2.hpp is not generated so move it as is to include/CL.

The old cl.hpp was the only header that was truly generated.
Commit the generated version and add automation to ensure that
the committed generated version matches the output of the
generation script with the committed template file.

Signed-off-by: Kevin Petit <kevin.petit@arm.com>