Test two additional defines and update descriptions in header
Newly tested defines:
* `CL_HPP_USE_CL_DEVICE_FISSION` (corrected from `CL_HPP_ENABLE_DEVICE_FISSION`)
* `CL_HPP_USE_CL_IMAGE2D_FROM_BUFFER_KHR`

The `CL_HPP_NO_STD_FOO` defines are not tested, because the header won't build
unless an alternative implementation is provided.  `The CL_HPP_USE_DX_INTEROP`
define is not tested because it does not work on platforms without DirectX
headers.

Issue #146 could make this irrelevant, but easy to test these defines in the
meantime.