Update clUpdateMutableCommandsKHR binding
In OpenCL-Docs PR https://github.com/KhronosGroup/OpenCL-Docs/pull/1045
the API for `cl_khr_command_buffer_mutable_dispatch` API
`clUpdateMutableCommandsKHR` changed in a breaking way.

When the headers update OpenCL-Headers PR https://github.com/KhronosGroup/OpenCL-Headers/pull/245
the bindings will break if they are not updated to reflect this
change.

Bindings updated in this PR to pass to arrays of templated length to
the C++ method. The underlying `.data()` pointers of these parameters can be
passed through to the OpenCL API.